avoid overcounting of non-try candidates
Before this CL, `if` statements with complex "then" branches or
non-nil `else` branches were counted (and interpreted as non-`try`
candidates) even though the `if` statement wasn't properly part
of an error check. Now, these cases are only counted if the `if`
statement can be part of an error check.

The output is now better organized into stats on `try` candidates
relative to all statements, and stats on reasons why valid error
checks are not candidates for `try`.

Removed the count of `return ... zero values, <err>` as it's not
providing extra information.